==> 3_VariancePartitioningFigure.R <==
setwd()

library(Hmsc)
library(pals)

#load Hmsc model generated by script 3.
load("models_thin_5000_samples_250_chains_4.Rdata")
m = models[[1]]

#Calculate variation partitioning (fractions of variance explained by the model's fixed and random effects per virus) and barplot the result:
  preds = computePredictedValues(m)
  VP = computeVariancePartitioning(m, group = attr(m$X[[1]], "assign"), groupnames = attr(terms(m$XFormula), "term.labels"))
  mycols = tol(nrow(VP$vals))

pdf("VarPart.pdf", width = 6, height = 5)
par(mar = c(5, 5, 3, 8.5), lwd = 0.1)
{
  barplot(VP$vals, ylim = c(0,1), cex.lab = 0.8, ylab = "Fractions of explained variance", cex.axis = 0.7, col = mycols, las=2, horiz=FALSE, cex.names = 0.6)
  legend("topright", inset = c(-0.48, 0), xpd=TRUE, legend = rev(c(all.vars(m$XFormula), "Random (plant)", "Random (population)")), fill = rev(mycols), bty = "n", cex = 0.8)
}
dev.off()

VPtable = round(VP$vals, 2)
