==> 5_BetaPlotFigure.R <==
setwd()

library(Hmsc)

#load Hmsc model generated by script 3.
load("models_thin_5000_samples_250_chains_4.Rdata")
m = models[[1]]

#Draw beta plot to illustrate the responses (positive, negative or neutral) of the five virus taxa to each model fixed effect covariate:
pdf("BetaPlot.pdf")
par(mar = c(6, 6, 3, 2), lwd = 0.1)
{
  m$covNames = c("Intercept", "Genotype 609_19", "Genotype 2818_6", "Genotype 4_13", "Closterovirus (ES)", "Betapartitivirus (ES)", "PILV (ES)", "Herbivory", "Plant size", "RNA n/a")
  postBeta = getPostEstimate(m, parName="Beta")
  plotBeta(m, post=postBeta, supportLevel = 0.95, param="Sign", colors = colorRampPalette(c("blue", "white", "yellow")),
           covNamesNumbers = c(TRUE,FALSE),
           spNamesNumbers=c(TRUE,FALSE),
           cex=c(0.6,0.6,0.8))
}
dev.off()


#Extract posterior beta means, lower and upper 95% credible intervals and medians. 

mpost = convertToCodaObject(m)
CIs = summary(mpost$Beta)
betatable = data.frame(CIs[[1]][, 1], CIs[[2]][, c(3, 1, 5)])
names(betatable) = c("mean", "median", "lower_CI_2.5%", "upper_CI_97.5%")
write.csv2(betatable, "Betamean_median_CIs.csv", row.names = T)
